+++ cli/main.cpp
@@ -277,6 +277,9 @@ static int mainGuarded(int argc, char** argv) {
   } case CmdRestore: {
     succ = restoreCrate(args);
     break;
+  } case CmdBackupPrune: {
+    succ = backupPruneCrate(args);
+    break;
   } case CmdReplicate: {
     succ = replicateCrate(args);
     break;